Déplace l'utilitaire d'aseptisation de valeur booléenne
