+++ src/utilitaires/aseptisation.js
@@ -1,0 +1,9 @@
+const valeurBooleenne = (valeur) => {
+  switch (valeur) {
+    case 'true': return true;
+    case 'false': return false;
+    default: return undefined;
+  }
+};
+
+module.exports = { valeurBooleenne };
